requirements.txt
python-telegram-bot
python-dateutil
python-dotenv
yfinance
requests